I edited HomeTrainerFragment.kt so it doesn't directly read from database insted it uses FirebaseManager.kt to read from database. FIT-14
